need to fix bug

+++ src/algorithm/merge_sort.h
@@ -7,6 +7,6 @@ class MergeSort {
   static void Sort(Sorter &sorter);
 
  private:
-  static void Recurse(int l, int r, Sorter &sorter);
-  static void InPlaceSort(int l, int r, int m, Sorter &sorter);
+  static void Recurse(size_t l, size_t r, Sorter &sorter);
+  static void InPlaceSort(size_t l, size_t r, size_t m, Sorter &sorter);
 };
