+++ src/algorithm/quick_sort.h
@@ -7,6 +7,6 @@ class QuickSort {
   static void Sort(Sorter &sorter);
 
  private:
-  static void Recurse(int l, int r, Sorter &sorter);
-  static size_t Partition(int l, int r, Sorter &sorter);
+  static void Recurse(size_t l, size_t r, Sorter &sorter);
+  static size_t Partition(size_t l, size_t r, Sorter &sorter);
 };
